+++ Sources/Sentry/SentryWatchdogTerminationScopeObserver.m
@@ -11,7 +11,7 @@
 @interface SentryWatchdogTerminationScopeObserver ()
 
 @property (nonatomic, strong) SentryWatchdogTerminationBreadcrumbProcessor *breadcrumbProcessor;
-@property (nonatomic, strong) SentryWatchdogTerminationContextProcessor *contextProcessor;
+@property (nonatomic, strong) SentryWatchdogTerminationAttributesProcessor *attributesProcessor;
 
 @end
 
@@ -19,11 +19,11 @@
 
 - (instancetype)
     initWithBreadcrumbProcessor:(SentryWatchdogTerminationBreadcrumbProcessor *)breadcrumbProcessor
-               contextProcessor:(SentryWatchdogTerminationContextProcessor *)contextProcessor;
+            attributesProcessor:(SentryWatchdogTerminationAttributesProcessor *)attributesProcessor;
 {
     if (self = [super init]) {
         self.breadcrumbProcessor = breadcrumbProcessor;
-        self.contextProcessor = contextProcessor;
+        self.attributesProcessor = attributesProcessor;
     }
 
     return self;
@@ -34,7 +34,7 @@
 - (void)clear
 {
     [self.breadcrumbProcessor clear];
-    [self.contextProcessor clear];
+    [self.attributesProcessor clear];
 }
 
 - (void)addSerializedBreadcrumb:(NSDictionary *)crumb
@@ -49,7 +49,7 @@
 
 - (void)setContext:(nullable NSDictionary<NSString *, id> *)context
 {
-    [self.contextProcessor setContext:context];
+    [self.attributesProcessor setContext:context];
 }
 
 - (void)setDist:(nullable NSString *)dist
@@ -84,7 +84,7 @@
 
 - (void)setUser:(nullable SentryUser *)user
 {
-    // Left blank on purpose
+    [self.attributesProcessor setUser:user];
 }
 
 - (void)setTraceContext:(nullable NSDictionary<NSString *, id> *)traceContext